backend/requirements.txt
fastapi
uvicorn
pymongo
pandas
pydantic
pyjwt
passlib[bcrypt]
jose
pycryptodome
python-jose

python-multipart
python-dotenv
torch

langchain
langchain_core
langchain_community
langchain_groq

feature_engine

# LIDA and LLMX dependencies
cohere
diskcache
google.auth
openai
pyyaml
tiktoken
typer
altair 
geopandas
kaleido 
matplotlib
matplotlib-venn
networkx
plotly
plotnine
pydantic
python-multipart
scipy
seaborn
statsmodels
typer
uvicorn
wordcloud
certifi>=2024.2.2